provisioning/cli/provisioning: move bmc logs commands to separate file
